Refactored Patient
